🚨 Fix: PR#20 수정 사항 #2 반영
