+++ src/pages/meditation/components/MeditationEndButton/MeditationEndButton.style.ts
@@ -1,0 +1,8 @@
+import styled from '@emotion/styled';
+
+export const EndButtonContainer = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-top: 50px;
+`;
