+++ src/pages/meditation/components/MeditationEndButton/index.ts
@@ -1,0 +1,3 @@
+import { MeditationEndButton } from './MeditationEndButton';
+
+export default MeditationEndButton;
