+++ src/pages/meditation/components/MeditationLabel/MeditationLabel.style.ts
@@ -1,0 +1,7 @@
+import styled from '@emotion/styled';
+
+export const StyledLabel = styled.div`
+  ${({ theme }) => theme.style.flexJustifyCenter};
+  color: ${({ theme }) => theme.color.white};
+  font-size: 19px;
+`;
