+++ src/pages/meditation/components/MeditationLabel/index.ts
@@ -1,0 +1,3 @@
+import { MeditationLabel } from './MeditationLabel';
+
+export default MeditationLabel;
