+++ src/pages/meditation/components/MeditationTimeSetter/index.ts
@@ -1,0 +1,3 @@
+import { MeditationTimeSetter } from './MeditationTimeSetter';
+
+export default MeditationTimeSetter;
