+++ src/pages/meditation/components/MeditationTimer/index.ts
@@ -1,0 +1,3 @@
+import { MeditationTimer } from './MeditationTimer';
+
+export default MeditationTimer;
