--- src/pages/meditation/constants/meditation.ts
@@ -1,11 +1,0 @@
-export const FIVE_MINUTES_IN_SECONDS = 300;
-export const COUNTER_BUTTON_SIZE = 90;
-export const BUTTON_TYPE_ADD = 'add';
-export const BUTTON_TYPE_SUB = 'sub';
-
-export const MEDITATION_NOT_STARTED = '명상 시간을 설정해주세요.';
-export const MEDITATION_ONGOING = '편안한 명상 되세요.';
-export const MEDITATION_ENDED = '명상이 종료되었습니다.';
-
-export const EVENT_NAME_MEDITATION_STARTED = 'meditation_started';
-export const EVENT_NAME_MEDITATION_ENDED = 'meditation_ended';
